chore: moves shell operations used for tests only to test/ pkg
