+++ e2e/completion/bash_completion_test.go
@@ -2,8 +2,8 @@ package completion_test
 
 import (
 	. "github.com/maistra/istio-workspace/e2e/infra"
-	"github.com/maistra/istio-workspace/pkg/shell"
 	"github.com/maistra/istio-workspace/test"
+	"github.com/maistra/istio-workspace/test/shell"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
